+++ client/src/App.jsx
@@ -2,8 +2,7 @@ import React, {useEffect} from 'react'
 import { Switch, Route } from 'react-router-dom'
 import { Register, Login, RegisterComplate, ForgotPassword } from './pages/auth/index'
 import { History, Wishlist, Password } from './pages/user/index'
-import UserRoute from './components/routes/UserRoute'
-import AdminRoute from './components/routes/AdminRoute'
+import {UserRoute,AdminRoute} from './components/routes/index'
 import { ToastContainer } from 'react-toastify'
 import 'react-toastify/dist/ReactToastify.css'
 import Home from './pages/Home'
