+++ client/src/components/nav/index.js
@@ -1,0 +1,3 @@
+export { default as AdminNav } from './AdminNav'
+export { default as Header } from './Header'
+export {default as UserNav} from './UserNav'
