+++ client/src/components/routes/index.js
@@ -1,0 +1,3 @@
+export { default as AdminRoute } from './AdminRoute'
+export { default as UserRoute } from './UserRoute'
+export { default as LoadingToRedirect} from './LoadingToRedirect'
