+++ client/src/pages/admin/category/CategoryUpdate.jsx
@@ -3,6 +3,7 @@ import AdminNav from '../../../components/nav/AdminNav'
 import { toast } from 'react-toastify'
 import { useSelector } from 'react-redux'
 import { getCategory, updateCategory } from '../../../functions/category'
+import CategoryForm from '../../../components/forms/CategoryForm'
 
 const CategoryUpdate = ({history, match}) => {
 
@@ -36,25 +37,6 @@ const CategoryUpdate = ({history, match}) => {
       })
   }
 
-  
-
-  const categoryForm = () => (
-    <form onSubmit={handleSubmit}>
-      <div className="form-group">
-        <label>Category Name</label>
-        <input
-          type="text"
-          className="form-control"
-          onChange={e => setName(e.target.value)}
-          value={name}
-          autoFocus
-          required
-        />
-        <br/>
-        <button className="btn btn-outline-primary">Update category</button>
-      </div>
-    </form>
-  )
   return (
     <div className ="container-fluid">
       <div className="row">
@@ -63,7 +45,12 @@ const CategoryUpdate = ({history, match}) => {
         </div>
         <div className="col">
           {loading ? <h4>Loading...</h4> : <h4>Update category</h4>}
-          {categoryForm()}
+          <CategoryForm
+            handleSubmit={handleSubmit}
+            name={name}
+            setName={setName}
+            text={"Update category"}
+          />
         </div>
       </div>
     </div>
